requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
anthropic
anyio
appnope
argon2-cffi
argon2-cffi-bindings
arrow
asttokens
async-lru
async-timeout
attrs
babel
beautifulsoup4
bleach
blis
boto3
botocore
cachetools
catalogue
cffi
click
cloudpathlib
comm
confection
contourpy
cryptography
cycler
cymem
dataclasses-json
debugpy
decorator
defusedxml
dill
distro
docopt
en_core_web_sm
et-xmlfile
exceptiongroup
executing
faiss-cpu
fastjsonschema
fonttools
fqdn
free_proxy
frozenlist
fsspec
git-filter-repo
googleapis-common-protos
graphviz
greenlet
groq
grpcio
grpcio-status
h11
hdfs
html2text
httpcore
httplib2
httpx
huggingface-hub
imbalanced-learn
ipykernel
ipython
ipywidgets
isoduration
jedi
jiter
jmespath
joblib
json5
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
jupyter
jupyter-console
jupyter-events
jupyter-lsp
jupyter_client
jupyter_core
jupyter_server
jupyter_server_terminals
jupyterlab
jupyterlab_pygments
jupyterlab_server
jupyterlab_widgets
keybert
kiwisolver
langchain
langchain-core
langchain-text-splitters
langcodes
langsmith
language_data
lxml
marisa-trie
markdown-it-py
marshmallow
matplotlib
matplotlib-inline
mdurl
minify_html
mistune
mkl-service
mpire
multidict
multiprocess
murmurhash
mypy-extensions
mysql-connector-python
nbclient
nbconvert
nbformat
nest-asyncio
nltk
notebook
notebook_shim
numpy
nvidia-nccl-cu12
oauth2client
oauthlib
ollama
openpyxl
orjson
outcome
overrides
packaging
pandas
pandocfilters
parso
patsy
pexpect
platformdirs
playwright
plotly
plotly-express
ply
preshed
prometheus_client
prompt_toolkit
propcache
proto-plus
protobuf
psutil
ptyprocess
pure_eval
pyasn1
pyasn1_modules
pycparser
pydantic
pydantic_core
pyee
Pygments
PyHDFS
PyMySQL
pyparsing
PyQt5
PyQt5-Qt5
PyQt5_sip
python-anticaptcha
python-dateutil
python-dotenv
python-json-logger
pytz
pyzmq
qtconsole
QtPy
referencing
regex
requests-oauthlib
requests-toolbelt
rfc3339-validator
rfc3986-validator
rich
rpds-py
rsa
s3transfer
safetensors
scikit-learn
scipy
seaborn
selenium
semchunk
Send2Trash
sentence-transformers
setuptools
shellingham
simplejson
sip
six
smart-open
sniffio
sortedcontainers
soupsieve
spacy
spacy-legacy
spacy-loggers
SQLAlchemy
srsly
stack-data
statsmodels
tenacity
terminado
textblob
thinc
threadpoolctl
tiktoken
tinycss2
tokenizers
tomli
torch
torchaudio
torchvision
tornado
tqdm
traitlets
transformers
trio
trio-websocket
triton
typer
types-python-dateutil
typing-inspect
tzdata
undetected-playwright
uri-template
uritemplate
wasabi
wcwidth
weasel
webcolors
webdriver-manager
webencodings
websocket-client
wheel
widgetsnbextension
wrapt
wsproto
xgboost
yarl
